# openapi.yaml
openapi: 3.0.0
info:
  title: Daily Journal API
  version: 1.0.0
  description: A simple API for managing daily journal entries and users with authentication
servers:
  - url: https://cse341-rlcp.onrender.com
    description: Production server
  - url: http://localhost:3000
    description: Local development server
  - url: http://localhost:3001
    description: Alternative local server (port 3001)
  - url: http://localhost:3002
    description: Alternative local server (port 3002)
  - url: http://localhost:3003
    description: Alternative local server (port 3003)
  - url: http://localhost:3004
    description: Alternative local server (port 3004)
  - url: http://localhost:3005
    description: Alternative local server (port 3005)
components:
  schemas:
...
    User:
      type: object
      required:
        - username
        - email
      properties:
        _id:
          type: string
          description: The auto-generated id of the user
        username:
          type: string
          description: The username of the user
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
          description: The email address of the user
        password:
          type: string
          description: The password of the user (min 6 characters)
          minLength: 6
        googleId:
          type: string
          description: Google OAuth ID for the user
        githubId:
          type: string
          description: GitHub OAuth ID for the user
        role:
          type: string
          enum: [user, admin]
          default: user
          description: The role of the user
        isActive:
          type: boolean
          default: true
          description: Whether the user account is active
        lastLogin:
          type: string
          format: date-time
          description: The last login timestamp
        preferences:
          type: object
          properties:
            theme:
              type: string
              enum: [light, dark]
              default: light
            notifications:
              type: boolean
              default: true
            language:
              type: string
              default: en
        createdAt:
          type: string
          format: date-time
          description: The creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: The last update timestamp
    AuthResponse:
      type: object
      properties:
        message:
          type: string
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
      required:
        - token
        - user
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    githubAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://github.com/login/oauth/authorize
          tokenUrl: https://github.com/login/oauth/access_token
          scopes:
            user:email: Read user email
            read:user: Read user profile
paths:
  /api/auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 30
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      summary: Login user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/auth/github/callback:
    get:
      summary: GitHub OAuth callback
      tags: [Auth]
      responses:
        '200':
          description: GitHub authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: GitHub authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/logout:
    post:
      summary: Logout user
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/me:
    get:
      summary: Get current user
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...